Refactor notifications
