+++ apps/wallet-mobile/src/features/Notifications/useCases/common/TransactionReceivedNotification.tsx
@@ -1,0 +1,181 @@
+import {useStrings} from './useStrings'
+import {useSelectedWallet} from '../../../WalletManager/common/hooks/useSelectedWallet'
+import {Balance, Notifications, Portfolio} from '@yoroi/types'
+import {NotificationItem} from './NotificationPopupItem'
+import * as React from 'react'
+import {StyleSheet, View} from 'react-native'
+import {useTheme} from '@yoroi/theme'
+import {Icon} from '../../../../components/Icon'
+import {TransactionInfo} from '../../../../yoroi-wallets/types/other'
+import {asQuantity, Quantities} from '../../../../yoroi-wallets/utils/utils'
+import {Token} from '../../../../yoroi-wallets/types/tokens'
+import {YoroiWallet} from '../../../../yoroi-wallets/cardano/types'
+
+export const getTransactionReceivedNotificationTitle = (
+  event: Notifications.Event,
+  strings: ReturnType<typeof useStrings>,
+  wallet: YoroiWallet,
+): string => {
+  if (event.trigger !== Notifications.Trigger.TransactionReceived) return ''
+
+  const isIntraWallet = wallet.transactions[event.metadata.txId]?.direction === 'SELF'
+  const isReceived = wallet.transactions[event.metadata.txId]?.direction === 'RECEIVED'
+  const isSent = wallet.transactions[event.metadata.txId]?.direction === 'SENT'
+
+  if (isIntraWallet) {
+    return strings.intraWalletTransactionSent
+  }
+
+  if (isReceived) {
+    const tx = wallet.transactions[event.metadata.txId]
+    if (tx === null) return ''
+    const details = getTransactionInfoDetails(tx, wallet.portfolioPrimaryTokenInfo)
+
+    return details.hasReceivedMultipleAssets
+      ? strings.multipleAssetsReceived
+      : `${formatAssets(
+          Quantities.format(details.firstAssetAmountReceived, details.firstReceivedAsset.denomination),
+          details.firstReceivedAsset.name,
+        )} ${strings.received}`
+  }
+
+  if (isSent) {
+    const tx = wallet.transactions[event.metadata.txId]
+    if (tx === null) return ''
+    const details = getTransactionInfoDetails(tx, wallet.portfolioPrimaryTokenInfo)
+
+    return details.hasSentMultipleAssets
+      ? strings.multipleAssetsSent
+      : `${formatAssets(
+          Quantities.format(details.firstAssetAmountSent, details.firstSentAsset.denomination),
+          details.firstSentAsset.name,
+        )} ${strings.sent}`
+  }
+
+  return ''
+}
+
+export const getTransactionReceivedNotificationIcon = (event: Notifications.Event, wallet: YoroiWallet) => {
+  if (event.trigger !== Notifications.Trigger.TransactionReceived) return null
+
+  const isIntraWallet = wallet.transactions[event.metadata.txId]?.direction === 'SELF'
+  const isReceived = wallet.transactions[event.metadata.txId]?.direction === 'RECEIVED'
+  const isSent = wallet.transactions[event.metadata.txId]?.direction === 'SENT'
+
+  if (isIntraWallet) {
+    return <Icon.Direction transactionDirection="SELF" />
+  }
+
+  if (isReceived) {
+    return <Icon.Direction transactionDirection="RECEIVED" />
+  }
+
+  if (isSent) {
+    return <Icon.Direction transactionDirection="SENT" />
+  }
+
+  return null
+}
+
+export const TransactionReceivedNotification = ({event}: {event: Notifications.Event}) => {
+  const strings = useStrings()
+  const {wallet} = useSelectedWallet()
+
+  if (event.trigger !== Notifications.Trigger.TransactionReceived) return null
+
+  return (
+    <NotificationItem
+      icon={<IconWrapper>{getTransactionReceivedNotificationIcon(event, wallet)}</IconWrapper>}
+      title={getTransactionReceivedNotificationTitle(event, strings, wallet)}
+      description={strings.tapToView}
+    />
+  )
+}
+const IconWrapper = ({children}: {children: React.ReactNode}) => {
+  const {styles, colors} = useStyles()
+  return <View style={[styles.icon, {backgroundColor: colors.iconBackground}]}>{children}</View>
+}
+
+const useStyles = () => {
+  const {atoms, color} = useTheme()
+  const styles = StyleSheet.create({
+    icon: {
+      width: 40,
+      height: 40,
+      borderRadius: 20,
+      ...atoms.align_center,
+      ...atoms.justify_center,
+    },
+  })
+
+  return {styles, colors: {iconColor: color.secondary_600, iconBackground: color.secondary_100}}
+}
+
+const sumTokenFromTxData = (
+  outputsOutputs: TransactionInfo['outputs'] | TransactionInfo['inputs'],
+  identifier: string,
+) => {
+  return outputsOutputs.reduce<Balance.Quantity>((acc, output) => {
+    const tokens = output.assets.filter((a) => a.identifier === identifier)
+    const quantities = tokens.map((t) => asQuantity(t.amount))
+    return Quantities.sum([acc, ...quantities])
+  }, Quantities.zero)
+}
+
+const findToken = (tokens: Token[], identifier: string, primaryTokenInfo: Portfolio.Token.Info) => {
+  if (identifier === primaryTokenInfo.id) {
+    return {name: primaryTokenInfo.name, denomination: primaryTokenInfo.decimals}
+  }
+
+  const token = tokens.find((t) => t.identifier === identifier)
+  const name = token?.metadata?.longName ?? token?.metadata?.ticker ?? identifier
+  const denomination = token?.metadata.numberOfDecimals ?? 0
+  return {name, denomination}
+}
+
+const sumPtFromOutputs = (outputs: TransactionInfo['outputs']) => {
+  return outputs.reduce<Balance.Quantity>((acc, output) => {
+    const amount = asQuantity(output.amount)
+    return Quantities.sum([acc, amount])
+  }, Quantities.zero)
+}
+
+const getTransactionInfoDetails = (info: TransactionInfo, primaryTokenInfo: Portfolio.Token.Info) => {
+  const ptReceived = sumPtFromOutputs(info.outputs)
+  const ptSent = sumPtFromOutputs(info.inputs)
+
+  const hasReceivedPt = !Quantities.isZero(ptReceived)
+  const hasSentPt = !Quantities.isZero(ptSent)
+
+  const assetsReceived = info.outputs.flatMap((o) => o.assets)
+  const assetsSent = info.inputs.flatMap((i) => i.assets)
+
+  const hasReceivedMultipleAssets = assetsReceived.length > 1 || (assetsReceived.length === 1 && hasReceivedPt)
+  const hasSentMultipleAssets = assetsSent.length > 1 || (assetsSent.length === 1 && hasSentPt)
+
+  const firstAssetIdReceived = assetsReceived[0]?.identifier ?? primaryTokenInfo.id
+  const firstAssetAmountReceived = hasReceivedPt ? ptReceived : sumTokenFromTxData(info.outputs, firstAssetIdReceived)
+  const firstReceivedAsset = hasReceivedPt
+    ? {name: primaryTokenInfo.name, denomination: primaryTokenInfo.decimals}
+    : findToken(Object.values(info.tokens), firstAssetIdReceived, primaryTokenInfo)
+
+  const firstAssetIdSent = assetsSent[0]?.identifier ?? primaryTokenInfo.id
+  const firstAssetAmountSent = hasSentPt ? ptSent : sumTokenFromTxData(info.inputs, firstAssetIdSent)
+  const firstSentAsset = hasSentPt
+    ? {name: primaryTokenInfo.name, denomination: primaryTokenInfo.decimals}
+    : findToken(Object.values(info.tokens), firstAssetIdSent, primaryTokenInfo)
+
+  return {
+    hasReceivedMultipleAssets,
+    hasSentMultipleAssets,
+    firstReceivedAsset,
+    firstSentAsset,
+    firstAssetAmountReceived,
+    firstAssetAmountSent,
+  }
+}
+
+const formatAssets = (quantity: string, name: string) => {
+  const text = `${quantity} ${name}`
+  return text.length > 15 ? `${text.slice(0, 15)}...` : text
+}
